feat(feat#2): split code to function named files
